+++ src/grafanarmadillo/paths.py
@@ -1,9 +1,9 @@
 """Tools for manipulating path-like objects into references to Grafana objects or file-safe paths."""
 from pathlib import Path
-from typing import List
+from typing import List, Sequence
 from urllib.parse import quote_plus, unquote_plus
 
-from grafanarmadillo.types import GrafanaPath
+from grafanarmadillo.types import GrafanaPath, PathLike
 
 
 class PathCodec:
@@ -12,7 +12,10 @@ class PathCodec:
 	@staticmethod
 	def encode_grafana(path: GrafanaPath) -> Path:
 		"""Encode a GrafanaPath."""
-		return PathCodec.encode([path.org, path.folder, path.name])
+		if path.org:
+			return PathCodec.encode([path.org, path.folder, path.name])
+		else:
+			return PathCodec.encode([path.folder, path.name])
 
 	@staticmethod
 	def encode(segments: List[str]) -> Path:
@@ -26,10 +29,20 @@ class PathCodec:
 		return quote_plus(segment)
 
 	@staticmethod
-	def decode_grafana(path: Path) -> GrafanaPath:
-		"""Decode a Path to a GrafanaPath."""
-		parts = PathCodec.decode(path)
+	def try_parse(o: PathLike) -> GrafanaPath:
+		"""Try to decode a pathlike object."""
+		if isinstance(o, GrafanaPath):
+			return o
+		elif isinstance(o, list):
+			return PathCodec.parse_grafana(o)
+		else:
+			path = Path(o)
+			parts = path.parts[1:] if path.is_absolute() else path.parts
+			return PathCodec.parse_grafana(parts)
 
+	@staticmethod
+	def parse_grafana(parts: Sequence[str]) -> GrafanaPath:
+		"""Assemble segments into an orderly GrafanaPath."""
 		if len(parts) == 3:
 			return GrafanaPath(org=parts[0], folder=parts[1], name=parts[2])
 		elif len(parts) == 2:
@@ -37,7 +50,7 @@ class PathCodec:
 		elif len(parts) == 1:
 			return GrafanaPath(folder="General", name=parts[0])
 		else:
-			raise ValueError(f"Grafana path has too many parts {path=} len={len(parts)}")
+			raise ValueError(f"Grafana path has too many parts {parts=} len={len(parts)}")
 
 	@staticmethod
 	def decode(path: Path) -> List[str]: