+++ tests/cli/test_cli.py
@@ -1,9 +1,10 @@
 import json
+from dataclasses import asdict
 
 import pytest
 from click.testing import CliRunner
 
-from grafanarmadillo.cmd import grafanarmadillo
+from grafanarmadillo.cmd import TemplatorOpts, grafanarmadillo
 from grafanarmadillo.find import Finder
 from tests.conftest import read_json_file, requires_alerting
 
@@ -106,6 +107,8 @@ def test_cli__import_alert(cli_config, rw_shared_grafana):
 			"prd",
 			"--mapping",
 			"file://tests/cli/mapping.json",
+			"--templator-extra-opts",
+			json.dumps(asdict(TemplatorOpts(remove_edit_metadata=True, resolve_alert_dashboarduid=True))),
 		]
 	)
 	assert result.exit_code == 0
@@ -115,6 +118,7 @@ def test_cli__import_alert(cli_config, rw_shared_grafana):
 	assert imported, "alert did not end up where we expected"
 	alert = gfn.alertingprovisioning.get_alertrule(imported["uid"])
 	assert "PRD" in alert["labels"].values()
+	assert "$$" not in alert["annotations"]["__dashboardUid__"]
 
 
 def test_cli__export_alert(cli_config, rw_shared_grafana, tmp_path):
@@ -139,10 +143,14 @@ def test_cli__export_alert(cli_config, rw_shared_grafana, tmp_path):
 			"template",
 			"--mapping",
 			"file://tests/cli/mapping.json",
+			"--templator-extra-opts",
+			json.dumps(asdict(TemplatorOpts(remove_edit_metadata=True, resolve_alert_dashboarduid=True))),
 		]
 	)
 	if result.exit_code != 0:
 		pytest.fail(result.output)
 
 	template = read_json_file(template_path)
-	assert "$tag1" in template["labels"].values()
+	assert "$tag1" in template["labels"].values(), "templating did not replace value with template"
+	assert "version" not in template, "templating did not remove edit metadata"
+	assert "$$" in template["annotations"]["__dashboardUid__"]